rename test_project to example_app
