+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.sourcepoint.test_project">
+    package="com.sourcepoint.example_app">
 
     <application
         android:allowBackup="false"
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.sourcepoint.test_project.MainActivity">
+        <activity android:name="com.sourcepoint.example_app.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
