+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sourcepoint.test_project.MainActivity">
+    tools:context="com.sourcepoint.example_app.MainActivity">
 
     <Button
         android:id="@+id/review_consents"
